SUPPORT2-161 Endpoint-ok létrehozása
Refactored some more code with the Scriptum_2023_IDEA code styling xml.

Added tests for invalid cases to controllers. (Achieved 90% coverage)

Refactored the deletion error handling.

Fixed a bug with multiple keyword savings. (It didn't check any of them for id's)

Added setMainTaskId to Task service and repository to set nullable value.

Added serialization and deserialization for UserStatusEnum for invalid controller test cases.

+++ src/main/java/org/example/service/impl/CompletedTasksServiceImpl.java
@@ -194,8 +194,16 @@ public class CompletedTasksServiceImpl implements CompletedTasksService
 	}
 
 	@Override
-	public void deleteCompletedTask(Long id)
+	public void deleteCompletedTask(Long id) throws ServiceException
 	{
+		CompletedTask found = getCompletedTaskById(id);
+		if (found == null)
+		{
+			throw new ServiceException(ServiceExceptionType.ILLEGAL_ID_ARGUMENT,
+				"Completed Task with given id does not exist!"
+			);
+		}
+
 		completedTasksRepository.deleteById(id);
 	}
